Provided Lagom 1.3 support
This includes support for the new scaladsl. Included in this work:

* Provide Play compile time DI components traits in Play 2.5 bundle lib
* Renamed lagom1-conductr-bundle-lib to lagom1-java-conductr-bundle-lib
* Moved ConductR components in lagom1-conductr-bundle-lib to new javadsl
  package
* Created new lagom1-scala-conductr-bundle-lib library
* Made both javadsl and scaladsl service locator implementations extend
  Lagom's CircuitBreakingServiceLocator, so that circuit breaking does
  get used when Lagom is deployed to production.

No attempt has been made to maintain binary compatibility in the javadsl
conductr bundle lib artifact. This is because it is actually an entirely
new artifact, it has a new name, and the classes are in a new package,
and so there can be no binary compatibility issues if both are loaded.
