--- lagom1-conductr-bundle-lib/src/main/resources/play/reference-overrides.conf
@@ -1,11 +1,0 @@
-# Hack to override some of Lagom's defaults
-
-# ConductR's config file loading logic will load this file with a higher
-# priority than reference.conf, but a lower priority than application.conf.
-# That allows ConductR to override Lagom's reference.conf (which can't happen
-# from in ConductR's own reference.conf), but still allow users to override
-# ConductR's settings in their application.conf.
-
-
-# Custom ConductR Application loader to load ConductR related configurations
-play.application.loader = "com.typesafe.conductr.bundlelib.lagom.ConductRApplicationLoader"
