--- lagom1-conductr-bundle-lib/src/main/resources/reference.conf
@@ -1,12 +1,0 @@
-########################################
-# Lagom ConductR Reference Config File #
-########################################
-
-# This is the reference config file that contains all the default settings.
-# Make your edits/overrides in your application.conf.
-
-
-# Enables the `ServiceLocatorModule` to register the `ConductRServiceLocator`.
-# The `ConductRServiceLocator` implements Lagom's ServiceLocator
-# by using the `LocationService` of the Play 2.5 conductr-bundle-lib.
-play.modules.enabled += "com.typesafe.conductr.bundlelib.lagom.ServiceLocatorModule"
