+++ lagom1-java-conductr-bundle-lib/src/main/scala/com/typesafe/conductr/bundlelib/lagom/javadsl/ConductRApplicationLoader.scala
@@ -1,0 +1,24 @@
+package com.typesafe.conductr.bundlelib.lagom.javadsl
+
+import com.typesafe.conductr.bundlelib.akka.{ Env => AkkaEnv }
+import com.typesafe.conductr.bundlelib.lagom.javadsl.{ Env => LagomEnv }
+import com.typesafe.conductr.bundlelib.play.api.{ Env => PlayEnv }
+import play.api._
+import play.api.inject.guice.{ GuiceApplicationBuilder, GuiceApplicationLoader }
+
+/**
+ * Including this class into a Lagom project will automatically set Lagom's
+ * configuration up from ConductR environment variables. Add the following
+ * to your application.conf in order to include it:
+ *
+ *    play.application.loader = "com.typesafe.conductr.bundlelib.lagom.javadsl.ConductRApplicationLoader"
+ */
+class ConductRApplicationLoader extends ApplicationLoader {
+  def load(context: ApplicationLoader.Context): Application = {
+    val conductRConfig = Configuration(AkkaEnv.asConfig) ++ Configuration(PlayEnv.asConfig) ++ Configuration(LagomEnv.asConfig)
+    val newConfig = context.initialConfiguration ++ conductRConfig
+    val newContext = context.copy(initialConfiguration = newConfig)
+    val prodEnv = Environment.simple(mode = Mode.Prod)
+    new GuiceApplicationLoader(new GuiceApplicationBuilder(environment = prodEnv)).load(newContext)
+  }
+}
