+++ project/Dependencies.scala
@@ -16,7 +16,7 @@ object Version {
   val play23             = "2.3.10"
   val play24             = "2.4.8"
   val play25             = "2.5.4"
-  val lagom1             = "1.2.0"
+  val lagom1             = "1.3.0-RC1"
   val reactiveStreams    = "1.0.0"
   val scala              = "2.11.8"
   val scalaTest          = "3.0.1"
@@ -35,7 +35,9 @@ object Library {
   val akka24Sse          = "de.heikoseeberger"      %% "akka-sse"                       % Version.akka24Sse
   val akka23ContribExtra = "com.typesafe.akka"      %% "akka-contrib-extra"             % Version.akka23ContribExtra
   val akka24ContribExtra = "com.typesafe.akka"      %% "akka-contrib-extra"             % Version.akka24ContribExtra
-  val lagom1Api          = "com.lightbend.lagom"    %% "lagom-javadsl-api"              % Version.lagom1
+  val lagom1ClientJavadsl  = "com.lightbend.lagom"  %% "lagom-javadsl-client"           % Version.lagom1
+  val lagom1ClientScaladsl = "com.lightbend.lagom"  %% "lagom-scaladsl-client"          % Version.lagom1
+  val lagom1ServerScaladsl = "com.lightbend.lagom"  %% "lagom-scaladsl-server"          % Version.lagom1
   val java8Compat        = "org.scala-lang.modules" % "scala-java8-compat_2.11"         % Version.java8Compat
   val reactiveStreams    = "org.reactivestreams"    % "reactive-streams"                % Version.reactiveStreams
   val junit              = "junit"                  %  "junit"                          % Version.junit
